Implements subtask execution framework
Adds a framework for executing subtasks within the mission handler.

This introduces a `SubtaskManager` class responsible for creating,
executing, and monitoring subtasks.  It also includes a base class
`SubtaskExecutorBase` for defining specific subtask behaviors.

Currently implements a `WaitSubtaskExecutor` as an example which blocks
for a specified duration.

The mission handler is updated to use the subtask manager and iterate
through lists of subtasks.